Book Exchange Project
Creation of the Website "Truequesina" which consists of a system that allows to manage and monitor the book exchange process at the University of Colima. During the development I was using technologies like HTML, CSS, Bootstrap and JavaScript for the Front-End and for the Backend I was using PHP together with my database in phpmyadmi.